Issue ID: CVE-2017-18349
Description: parseObject in Fastjson before 1.2.25, as used in FastjsonEngine in Pippo 1.11.0 and other products, allows remote attackers to execute arbitrary code via a crafted JSON request, as demonstrated by a crafted rmi:// URI in the dataSourceName field of HTTP POST data to the Pippo /json URI, which is mishandled in AjaxApplication.java.

Fixed Branch: This vulnerability has been patched since 1.2.25.

Rootcause:
The root cause of the vulnerability is still the autotype function of fastjson, which can manually specify a well-designed class when deserializing to achieve remote command execution

Steps to reproduce:
- The attacker prepares rmi services and web services and injects rmi absolute path into the lookup method.
- The victim's JNDI interface points to the attacker to control the rmi server.
- The JNDI interface loads malicious code from the attacker's web server and executes the constructor to cause RCE.

Fix:
- The patching method will be com.alibaba.fastjson.parser.DefaultJSONParser#parseObject
- Call in method com.alibaba.fastjson.parser.ParserConfig#checkAutoType To check whether the class we pass in is in the blacklist, that is, TypeUtils.loadClass is replaced with the checkAutoType() function
- loadClass will be called only after passing the whitelist verification
- However, both whitelist and blacklist are used to restrict the deserialized classes.
- Only when the whitelist does not pass will blacklist be judged, which means that the whitelist does not really play the role of whitelist